Bootstrappers for terminal, zsh and others

+++ bootstrappers/bootstrap-jdk.sh
@@ -15,26 +15,51 @@ info() {
 
 main() {
   init_logger
-  info "Configuring jenv"; # --------------------------------------------------------------------------
+  info "This bootstrapper is going to download and install Ruby and rbenv via Homebrew"; # -------------------------------------------------------------------------
+  info "Would you like to continue?"
+  while true; do
+    read -p "$prefix Enter [y|n]:" answer
+    case $answer in
+      [Yy])
+        if [ ! type -t brew ]
+        then
+          if [[ ! -e /opt/homebrew/bin/brew ]]
+            info "Cannot run this bootstrapper. Install required dependency first: Homebrew."
+          else
+            eval "$(/opt/homebrew/bin/brew shellenv)"
 
-  brew install openjdk@11
+            brew install openjdk@11
+            brew install jenv
 
-  brew install jenv
+            echo '### jEnv' >> $HOME/env.sh
+            echo 'export PATH="$HOME/.jenv/bin:$PATH"' >> $HOME/env.sh
+            echo 'eval "$(jenv init -)"' >> $HOME/env.sh
 
-  # If the ~/.jenv/versions folder is empty, add all of the current JDK installations into jenv.
-  if [ -z "$(find "${HOME}/.jenv/versions" -mindepth 1 -maxdepth 1 -print -quit 2> /dev/null)" ]; then
-    while IFS= read -r -d $'\n' jdk; do
-      if [ -d "${jdk}/Contents/Home/bin" ]; then
-        jenv add "${jdk}/Contents/Home";
-      fi;
-    done < <(find /Library/Java/JavaVirtualMachines -type d -maxdepth 1 -mindepth 1 -print);
-  fi;
+            source $HOME/env.sh
 
-  # Default to Java 11.
-  if [ ! -r "${HOME}/.jenv/version" ]; then
-    jenv global 11;
-    jenv rehash;
-  fi;
+            # If the ~/.jenv/versions folder is empty, add all of the current JDK installations into jenv.
+            if [ -z "$(find "${HOME}/.jenv/versions" -mindepth 1 -maxdepth 1 -print -quit 2> /dev/null)" ]; then
+              while IFS= read -r -d $'\n' jdk; do
+                if [ -d "${jdk}/Contents/Home/bin" ]; then
+                  jenv add "${jdk}/Contents/Home";
+                fi;
+              done < <(find /Library/Java/JavaVirtualMachines -type d -maxdepth 1 -mindepth 1 -print);
+            fi;
+
+            # Default to Java 11.
+            if [ ! -r "${HOME}/.jenv/version" ]; then
+              jenv global 11;
+              jenv rehash;
+            fi;
+          fi
+        fi
+      ;;
+      [Nn]) 
+        info "Skipping...";
+        break;
+      ;;
+    esac
+  done
 }
 
 main "$@";
